+++ package.json
@@ -18,7 +18,6 @@
   },
   "devDependencies": {
     "@biomejs/biome": "^2.2.7",
-    "arktype": "^2.1.23",
     "bumpp": "^9.11.1",
     "turbo": "^2.5.8",
     "typescript": "catalog:",
