+++ packages/core/package.json
@@ -18,7 +18,6 @@
     "clean": "rm -rf ./node_modules"
   },
   "devDependencies": {
-    "@better-call/node": "workspace:^",
     "arktype": "^2.0.0",
     "tsdown": "catalog:",
     "type-fest": "^4.23.0",
